feat: 스프링 카페 1-3단계 구현

+++ src/main/java/com/kakao/cafe/controller/BoardController.java
@@ -83,7 +83,7 @@ public class BoardController {
 
     @DeleteMapping("/delete/{articleId}/{commentId}")
     public String deleteComment(@PathVariable("articleId") long articleId,
-                                @PathVariable("commentId") int commentId) {
+                                @PathVariable("commentId") long commentId) {
         boardService.deleteComment(articleId, commentId);
 
         return "redirect:/board/view/" + articleId;
